testing for consensus classifiers
